+++ src/main/java/roomescape/controller/helper/LoginMember.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.member;
+package roomescape.controller.helper;
 
 import roomescape.domain.member.Role;
