+++ src/main/java/roomescape/controller/reservation/ReservationApiController.java
@@ -15,9 +15,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.controller.dto.AdminReservationRequest;
 import roomescape.controller.dto.MemberReservationRequest;
-import roomescape.controller.resolver.AuthenticationPrincipal;
+import roomescape.controller.helper.AuthenticationPrincipal;
+import roomescape.controller.helper.LoginMember;
 import roomescape.service.ReservationService;
-import roomescape.service.dto.member.LoginMember;
 import roomescape.service.dto.reservation.ReservationCreate;
 import roomescape.service.dto.reservation.ReservationResponse;
 import roomescape.service.dto.reservation.ReservationSearchParams;
