+++ src/main/java/roomescape/service/dto/theme/PopularThemeRequest.java
@@ -3,7 +3,7 @@ package roomescape.service.dto.theme;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import roomescape.service.dto.validator.DateFormatConstraint;
+import roomescape.controller.helper.validator.DateFormatConstraint;
 
 public class PopularThemeRequest {
