+++ src/main/java/roomescape/service/dto/time/AvailableTimeRequest.java
@@ -2,7 +2,7 @@ package roomescape.service.dto.time;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
-import roomescape.service.dto.validator.DateFormatConstraint;
+import roomescape.controller.helper.validator.DateFormatConstraint;
 
 public class AvailableTimeRequest {
