+++ src/test/java/roomescape/utils/CookieUtilsTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
+import roomescape.global.CookieUtils;
 
 class CookieUtilsTest {
